Address review: drop shared abstract check and align S9354 impact.
S9148 and S9354 now share only ComparisonMethodUtils instead of an
abstract check class. GeneratedCheckListTest no longer needs a
blacklist entry, and RELIABILITY is MEDIUM to match Major severity.
