+++ java-checks/src/main/java/org/sonar/java/checks/FloatingPointComparisonCheck.java
@@ -16,27 +16,46 @@
  */
 package org.sonar.java.checks;
 
+import java.util.Arrays;
+import java.util.List;
 import org.sonar.check.Rule;
+import org.sonar.java.checks.helpers.ComparisonMethodUtils;
+import org.sonar.plugins.java.api.IssuableSubscriptionVisitor;
 import org.sonar.plugins.java.api.semantic.Type;
+import org.sonar.plugins.java.api.tree.BaseTreeVisitor;
 import org.sonar.plugins.java.api.tree.BinaryExpressionTree;
+import org.sonar.plugins.java.api.tree.ClassTree;
 import org.sonar.plugins.java.api.tree.ExpressionTree;
 import org.sonar.plugins.java.api.tree.LambdaExpressionTree;
 import org.sonar.plugins.java.api.tree.MethodTree;
 import org.sonar.plugins.java.api.tree.Tree;
 
 @Rule(key = "S9148")
-public class FloatingPointComparisonCheck extends AbstractComparisonMethodCheck {
+public class FloatingPointComparisonCheck extends IssuableSubscriptionVisitor {
 
   private static final String MESSAGE = "Use \"Double.compare\" or \"Float.compare\" to compare floating-point values.";
 
   @Override
-  void visitComparisonMethod(MethodTree methodTree) {
-    methodTree.block().accept(new FloatingPointComparisonVisitor());
+  public List<Tree.Kind> nodesToVisit() {
+    return Arrays.asList(Tree.Kind.METHOD, Tree.Kind.LAMBDA_EXPRESSION);
   }
 
   @Override
-  void visitComparatorLambda(LambdaExpressionTree lambda) {
-    lambda.body().accept(new FloatingPointComparisonVisitor());
+  public void visitNode(Tree tree) {
+    if (context.getSemanticModel() == null) {
+      return;
+    }
+    if (tree.is(Tree.Kind.METHOD)) {
+      MethodTree methodTree = (MethodTree) tree;
+      if (ComparisonMethodUtils.isCompareMethod(methodTree)) {
+        methodTree.block().accept(new FloatingPointComparisonVisitor());
+      }
+    } else {
+      LambdaExpressionTree lambda = (LambdaExpressionTree) tree;
+      if (ComparisonMethodUtils.isComparatorLambda(lambda)) {
+        lambda.body().accept(new FloatingPointComparisonVisitor());
+      }
+    }
   }
 
   private static boolean hasFloatingType(ExpressionTree tree) {
@@ -44,7 +63,7 @@ public class FloatingPointComparisonCheck extends AbstractComparisonMethodCheck 
       || tree.symbolType().isPrimitive(Type.Primitives.DOUBLE);
   }
 
-  private class FloatingPointComparisonVisitor extends IgnoreNestedTypesVisitor {
+  private class FloatingPointComparisonVisitor extends BaseTreeVisitor {
 
     @Override
     public void visitBinaryExpression(BinaryExpressionTree tree) {
@@ -58,6 +77,16 @@ public class FloatingPointComparisonCheck extends AbstractComparisonMethodCheck 
 
     private boolean hasFloatingOperand(BinaryExpressionTree tree) {
       return hasFloatingType(tree.leftOperand()) || hasFloatingType(tree.rightOperand());
+    }
+
+    @Override
+    public void visitClass(ClassTree tree) {
+      // Do not visit inner classes
+    }
+
+    @Override
+    public void visitLambdaExpression(LambdaExpressionTree tree) {
+      // Do not visit nested lambdas
     }
   }
 }
