+++ sonar-java-plugin/src/main/resources/org/sonar/l10n/java/rules/java/S9354.json
@@ -16,7 +16,7 @@
   "quickfix": "targeted",
   "code": {
     "impacts": {
-      "RELIABILITY": "HIGH"
+      "RELIABILITY": "MEDIUM"
     },
     "attribute": "LOGICAL"
   }
